:recycle:refactor: refactor book search page 사용하지 않는 코드 삭제 Related to: #269
