+++ src/features/BookDetailPage/slice/readerStatsSlice.ts
@@ -1,5 +1,32 @@
-import { ReaderStats } from '@components/BookDetailPage/types';
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+
+export interface ReaderStats {
+  isbn: string;
+  totalCollectors: number;
+  demographics: {
+    gender: {
+      male: {
+        '10s': number;
+        '20s': number;
+        '30s': number;
+        '40s': number;
+        '50s': number;
+        '60plus': number;
+        unknown: number;
+      };
+      female: {
+        '10s': number;
+        '20s': number;
+        '30s': number;
+        '40s': number;
+        '50s': number;
+        '60plus': number;
+        unknown: number;
+      };
+      unknown: number;
+    };
+  };
+}
 
 export const defaultReaderStats: ReaderStats = {
   isbn: '',
