+++ src/features/BookSearchPage/types/types.ts
@@ -46,14 +46,6 @@ export interface BookResponse {
   }>;
 }
 
-// 요청 파라미터 타입 정의
-export interface SearchBooksParams {
-  query: string; // 사용자 검색어
-  page: number; // 페이지 넘버
-  sort: string; // 정렬 기준
-  maxResults?: number;
-}
-
 // 상품 조회 관련 인터페이스
 export interface RatingInfoResponse {
   item: Array<{
